+++ src/switchable/autoplay.js
@@ -2,7 +2,7 @@
  * Switchable Autoplay Plugin
  * @creator  玉伯<lifesinger@gmail.com>
  */
-KISSY.add('switchable/autoplay', function(S, Event,Switchable,undefined) {
+KISSY.add('switchable/autoplay', function(S, Event, Switchable, undefined) {
 
 
     /**
@@ -24,6 +24,7 @@ KISSY.add('switchable/autoplay', function(S, Event,Switchable,undefined) {
         name: 'autoplay',
 
         init: function(host) {
+
             var cfg = host.config, interval = cfg.interval * 1000, timer;
             if (!cfg.autoplay) return;
 
@@ -31,11 +32,9 @@ KISSY.add('switchable/autoplay', function(S, Event,Switchable,undefined) {
             if (cfg.pauseOnHover) {
                 Event.on(host.container, 'mouseenter', function() {
                     host.stop();
-                    host.paused = true; // paused 可以让外部知道 autoplay 的当前状态
                 });
                 Event.on(host.container, 'mouseleave', function() {
-                    host.paused = false;
-                    startAutoplay();
+                    host.start();
                 });
             }
 
@@ -43,7 +42,6 @@ KISSY.add('switchable/autoplay', function(S, Event,Switchable,undefined) {
                 // 设置自动播放
                 timer = S.later(function() {
                     if (host.paused) return;
-
                     // 自动播放默认 forward（不提供配置），这样可以保证 circular 在临界点正确切换
                     host.switchTo(host.activeIndex < host.length - 1 ? host.activeIndex + 1 : 0, 'forward');
                 }, interval, true);
@@ -58,8 +56,21 @@ KISSY.add('switchable/autoplay', function(S, Event,Switchable,undefined) {
                     timer.cancel();
                     timer = undefined;
                 }
-            }
+                host.paused = true; // paused 可以让外部知道 autoplay 的当前状态
+                S.log("stop");
+            };
+
+            host.start = function() {
+                if (timer) {
+                    timer.cancel();
+                    timer = undefined;
+                }
+                host.paused = false;
+                S.log("start");
+                startAutoplay();
+
+            };
         }
     });
     return Switchable;
-}, { requires:["event","switchable/base"]});
+}, { requires:["event","./base"]});